requirements.txt
python-dotenv==1.0.0
selenium==3.141.0
urllib3==1.26.4
